backend/requirements.txt
Flask>=2.3.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
tensorflow>=2.13.0
deepface>=0.0.93
gTTS>=2.5.0
moviepy>=1.0.3
mysql-connector-python>=8.0.0
